% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinlsac-package.R
\docType{package}
\name{joinlsac-package}
\alias{joinlsac}
\alias{joinlsac-package}
\title{joinlsac: Joins LSAC Data Files for Further Analysis}
\description{
A tool to read and join data files from the Longitudinal Study of Australian Children (LSAC) for further analysis.
}
\author{
\strong{Maintainer}: Taren Sanders \email{taren.sanders@acu.edu.au}

}
\keyword{internal}
