% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinsurveys.R
\name{joinsurveys}
\alias{joinsurveys}
\title{Join LSAC Survey Data}
\usage{
joinsurveys(datadir, cohorts = c("b", "k"), ages = "all", vars)
}
\arguments{
\item{datadir}{The file path to the LSAC dataset.}

\item{cohorts}{A character vector of the cohorts to include (i.e., either "b"
or "k", or bothks).}

\item{ages}{A character vector of the ages to include (e.g., "0", "2"...).
You can also specify "all" to include all available ages.}

\item{vars}{(Not implemented) A character vector of variables to include in
the output. Note that this is optional and that all variables are returned
by default.}
}
\value{
A dataframe in long format.
}
\description{
Reads and joins the survey data from the LSAC dataset.
}
\examples{
# ADD_EXAMPLES_HERE
}
